reconfigure mcp server to play a bit nicer with uvx
to ensure compatibility with uvx, I changed the following:

1. Merged the two servers into one server so that only one installation from opentools was needed
2. Modified the __init__.py & __main__.py so that you can import and/or run the program needed
3. Modified other ancillary files such as sample_claude_desktop_config.json & pyproject.toml & bumpversion.cfg & .github/workflows/main.yaml
4. I moved stuff out of src & into artemis_mcp because some import stuff didn't seem to be playing nice when I ran the server locally using uvx.

After this, I think you may need to bump the version & publish on pypi for "ARTEMIS_API_KEY=abc123 uvx artemis-mcp" to work properly
