+++ .bumpversion.cfg
@@ -8,6 +8,6 @@ tag_name = {new_version}
 search = version = "{current_version}"
 replace = version = "{new_version}"
 
-[bumpversion:file:src/__init__.py]
+[bumpversion:file:artemis_mcp/__init__.py]
 search = __version__ = "{current_version}"
 replace = __version__ = "{new_version}"
