+++ .github/workflows/main.yml
@@ -49,7 +49,7 @@ jobs:
         run: |
           git config --local user.email "github-actions[bot]@users.noreply.github.com"
           git config --local user.name "github-actions[bot]"
-          git add pyproject.toml src/__init__.py .bumpversion.cfg
+          git add pyproject.toml artemis_mcp/__init__.py .bumpversion.cfg
           git commit -m "Bump version [skip ci]"
       - name: Push changes
         uses: ad-m/github-push-action@master
