+++ artemis_mcp/__init__.py
@@ -1,0 +1,13 @@
+__version__ = "0.1.5"
+
+from . import server
+
+
+def main():
+    server.main()
+
+
+__all__ = [
+    "server",
+    "main",
+]
